sdfsdf

+++ auth-service/src/main/java/com/emm/authservice/dtos/AuthUserDto.java
@@ -6,12 +6,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class AuthUserDto {
-    private Integer id;
+    private Long id;
     private String userName;
     private String password;
 }
