+++ auth-service/src/main/java/com/emm/authservice/repository/AuthUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AuthUserRepository extends JpaRepository<AuthUser,Integer> {
+public interface AuthUserRepository extends JpaRepository<AuthUser,Long> {
 
     Optional<AuthUser> findByUserName(String userName);
